get_agent_news/pyproject.toml
[project]
name = "get_agent_news"
version = "0.1.0"
description = "Weekly AI Agent/LLM news fetcher"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "tenacity>=8.2",
    "python-dateutil>=2.9",
    "PyYAML>=6.0",
]

[tool.uv]
# uv will infer from requires-python and dependencies

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
